saved_models/convert.py
from keras.models import load_model
import tensorflow as tf

model = load_model("catdog.h5")


converter = tf.lite.TFLiteConverter.from_keras_model(model)
tflite_model = converter.convert()

print("model converted")

# Save the model.
with open('model.tflite', 'wb') as f:
  f.write(tflite_model)
  
